fix command-line argument parsing segfault (leading space)
The bug is triggered by passing a command-line argument containing a leading
space character or by passing an empty leading command-line argument with an
additional argument after it.

Several of the newly added tests currently only check whether the test crashes
and not the command-line arguments left over after UTF processing. This is so
because the current UTF implementation does not seem to handle arguments with
embedded spaces correctly and that seems like a separate issue.