Replace BOOST_TEST_USE_DEPRECATED_TIMER macro with --deprecated_timer_format
Remove the macro BOOST_TEST_USE_DEPRECATED_TIMER and replace it with a
runtime cla --deprecated_timer_format. Additionally change <TestingTime>
to write the testing time in microseconds and add <CpuTime> giving the
time in nanoseconds. Existing scripts and parsers that handle the XML
output should not be affected by this change.