Datasets based on std::initializer_list fixes
Since the previous changes, the elements of the std::initializer_list
get destroyed before the test tree initializes properly. The changes
are

- not relying on a moved std::initializer_list for storing the values
of the dataset. Those are moved to a std::vector instead if the
move constructor is noexcept as expected by std::vector,
- providing a template parameter pack for the same type of constructs
in place of the std::initializer_list

Some of the changes are also related to the fact that VS2013 does not
handle properly the std::initializer_list vs template parameter pack.