Master test suite name setup prior to all other setup
This is a simple call to a global setting that can be performed

- in an idempotent manner
- prior to any other setup

Since some loggers rely on the master test suite (generation of the
output files), this setting should precede any other. In particular,
it should not be delayed until the user `init_unit_test` is called. A global object fits this purpose.