Removing Travis and Appveyor from the front README and the code.
Appveyor /NEVER/ worked as any of the builds is taking up too much time.
This is quite confusing for the users/developers.

Travis is dead slow and it is virtually impossible to get any build
status in reasonable time. The free tier is now discontinued.

Codecov: never got a chance to get something good out of it.

Those builds were pushed to Boost.Test without any serious commitment
to maintain them on the long run.